requirements.txt
pandas
numpy
faiss-cpu
google-generativeai
py2neo
sentence-transformers
fastapi
uvicorn[standard]
openpyxl
